MAR-848 Created the getRoutesList helper
